[Tests-only] Refactor functions from CommandLine trait to contexts and helpers
